requirements.txt
django==4.1.7
environs==9.5.0
dj-database-url==1.2.0
postgres
pillow

python-telegram-bot==13.12
redis
requests
